[PC-253] page -> graph 이름 변경
