+++ feature/auth/src/main/java/com/puzzle/auth/graph/main/contract/AuthIntent.kt
@@ -1,0 +1,3 @@
+package com.puzzle.auth.graph.main.contract
+
+class AuthIntent
