+++ feature/auth/src/main/java/com/puzzle/auth/graph/main/contract/AuthSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.auth.graph.main.contract
+
+class AuthSideEffect
