+++ feature/auth/src/main/java/com/puzzle/auth/graph/main/contract/AuthState.kt
@@ -1,0 +1,7 @@
+package com.puzzle.auth.graph.main.contract
+
+import com.airbnb.mvrx.MavericksState
+
+data class AuthState(
+    val a: Boolean = false,
+) : MavericksState
