+++ feature/auth/src/main/java/com/puzzle/auth/graph/registration/contract/AuthRegistrationIntent.kt
@@ -1,0 +1,3 @@
+package com.puzzle.auth.graph.registration.contract
+
+sealed class AuthRegistrationIntent
