+++ feature/auth/src/main/java/com/puzzle/auth/graph/registration/contract/AuthRegistrationSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.auth.graph.registration.contract
+
+sealed class AuthRegistrationSideEffect
