+++ feature/auth/src/main/java/com/puzzle/auth/graph/verification/contract/AuthVerficationState.kt
@@ -1,0 +1,7 @@
+package com.puzzle.auth.graph.verification.contract
+
+import com.airbnb.mvrx.MavericksState
+
+data class AuthVerficationState(
+    val a: Boolean = false,
+) : MavericksState
