+++ feature/auth/src/main/java/com/puzzle/auth/graph/verification/contract/AuthVerificationSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.auth.graph.verification.contract
+
+sealed class AuthVerificationSideEffect
