--- feature/auth/src/main/java/com/puzzle/auth/page/main/contract/AuthIntent.kt
@@ -1,4 +1,0 @@
-package com.puzzle.auth.page.main.contract
-
-class AuthIntent {
-}
