--- feature/auth/src/main/java/com/puzzle/auth/page/main/contract/AuthState.kt
@@ -1,7 +1,0 @@
-package com.puzzle.auth.page.main.contract
-
-import com.airbnb.mvrx.MavericksState
-
-data class AuthState(
-    val a: Boolean,
-) : MavericksState
