--- feature/auth/src/main/java/com/puzzle/auth/page/registration/AuthRegistrationScreen.kt
@@ -1,13 +1,0 @@
-package com.puzzle.auth.page.registration
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import com.airbnb.mvrx.compose.collectAsState
-import com.airbnb.mvrx.compose.mavericksViewModel
-
-@Composable
-internal fun AuthRegistrationScreen(
-    viewModel: AuthRegistrationViewModel = mavericksViewModel()
-) {
-    val state by viewModel.collectAsState()
-}
