--- feature/auth/src/main/java/com/puzzle/auth/page/verification/contract/AuthVerficationState.kt
@@ -1,7 +1,0 @@
-package com.puzzle.auth.page.verification.contract
-
-import com.airbnb.mvrx.MavericksState
-
-data class AuthVerficationState(
-    val a: Boolean,
-) : MavericksState
