--- feature/auth/src/main/java/com/puzzle/auth/page/verification/contract/AuthVerificationIntent.kt
@@ -1,4 +1,0 @@
-package com.puzzle.auth.page.verification.contract
-
-sealed class AuthVerificationIntent {
-}
