--- feature/auth/src/main/java/com/puzzle/auth/page/verification/contract/AuthVerificationSideEffect.kt
@@ -1,4 +1,0 @@
-package com.puzzle.auth.page.verification.contract
-
-sealed class AuthVerificationSideEffect {
-}
