+++ feature/matching/src/main/java/com/puzzle/matching/graph/detail/contract/MatchingDetailSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.matching.graph.detail.contract
+
+sealed class MatchingDetailSideEffect
