+++ feature/matching/src/main/java/com/puzzle/matching/graph/main/contract/MatchingIntent.kt
@@ -1,0 +1,5 @@
+package com.puzzle.matching.graph.main.contract
+
+sealed class MatchingIntent {
+    data object NavigateToMatchingDetail : MatchingIntent()
+}
