+++ feature/matching/src/main/java/com/puzzle/matching/graph/main/contract/MatchingSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.matching.graph.main.contract
+
+sealed class MatchingSideEffect
