--- feature/matching/src/main/java/com/puzzle/matching/page/detail/contract/MatchingDetailIntent.kt
@@ -1,8 +1,0 @@
-package com.puzzle.matching.page.detail.contract
-
-sealed class MatchingDetailIntent {
-    data object OnMatchingDetailCloseClick : MatchingDetailIntent()
-    data object OnPreviousPageClick : MatchingDetailIntent()
-    data object OnNextPageClick : MatchingDetailIntent()
-    data object OnMoreClick : MatchingDetailIntent()
-}
