--- feature/matching/src/main/java/com/puzzle/matching/page/detail/contract/MatchingDetailSideEffect.kt
@@ -1,3 +1,0 @@
-package com.puzzle.matching.page.detail.contract
-
-sealed class MatchingDetailSideEffect
