--- feature/matching/src/main/java/com/puzzle/matching/page/main/contract/MatchingSideEffect.kt
@@ -1,3 +1,0 @@
-package com.puzzle.matching.page.main.contract
-
-sealed class MatchingSideEffect
